Migrated addNewCoordinate func to GPSMapBranch
