+++ Map Map/Models/GPSMap.swift
@@ -13,8 +13,6 @@ import MapKit
 @objc(GPSMap)
 public class GPSMap: NSManagedObject { 
     /// A simple getter for GPS Map Coordinates.
-    var unwrappedConnections: [GPSMapCoordinateConnection] {
-        return (self.connections?.array as? [GPSMapCoordinateConnection] ?? []).filter { $0.end != nil }
     }
     
     /// Track how this GPS map is being edited.
@@ -29,37 +27,6 @@ public class GPSMap: NSManagedObject {
     var unwrappedEditing: EditingState {
         get { return EditingState(rawValue: self.editing) ?? .settingUp }
         set(newValue) { editing = newValue.rawValue }
-    }
-    
-    /// Add a new coordinate to this GPS Map.
-    /// - Parameter clLocation: Location to add.
-    /// - Returns: Added GPSMapCoordinate if it was successful.
-    public func addNewCoordinate(clLocation: CLLocation) -> GPSMapCoordinate? {
-        guard let moc = self.managedObjectContext
-        else { return nil }
-        let truncAlt = Int16(clLocation.altitude) // Track min/max altitude
-        let newCoordinate = GPSMapCoordinate(location: clLocation, moc: moc)
-        
-        if let lastConnection = self.unwrappedConnections.last {
-            if self.heightMax < truncAlt { self.heightMax = truncAlt }
-            else if self.heightMin > truncAlt { self.heightMin = truncAlt }
-            
-            if let startCoordinate = lastConnection.end { // Create a new connection using previous end as new start
-                self.addToConnections(GPSMapCoordinateConnection(start: startCoordinate, end: newCoordinate, context: moc))
-                self.distance += Int32(clLocation.distance(from: startCoordinate.clLocation))
-            }
-            else if let startCoordinate = lastConnection.start { // Use the existing connection with current as new end
-                lastConnection.end = newCoordinate
-                self.distance += Int32(clLocation.distance(from: startCoordinate.clLocation))
-            }
-            return newCoordinate
-        }
-        
-        self.heightMax = truncAlt
-        self.heightMin = truncAlt
-        // Create new connection with current as start
-        self.addToConnections(GPSMapCoordinateConnection(start: newCoordinate, context: moc))
-        return newCoordinate
     }
 }
